System profiler test

+++ solardatatools/algorithms/capacity_change.py
@@ -71,7 +71,7 @@ class CapacityChange:
             self.labels = None
         else:
             if dbscan_min_samples == "auto":
-                dbscan_min_samples = max(0.1 * len(s1), 3)
+                dbscan_min_samples = int(max(0.1 * len(s1), 3))
             db = DBSCAN(eps=dbscan_eps, min_samples=dbscan_min_samples).fit(
                 s1[:, np.newaxis]
             )
